Fixed project formatting
