--- sEndBoatFixer/src/main/java/uk/stuqad/sendboatfixer/Listeners/OnEntityTeleportEvent.java
@@ -1,17 +1,0 @@
-package uk.stuqad.sendboatfixer.Listeners;
-
-import org.bukkit.World;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.entity.EntityTeleportEvent;
-
-public class OnEntityTeleportEvent implements Listener {
-
-    @EventHandler
-    public void onEntityTeleportEvent(EntityTeleportEvent event) {
-        if (event.getEntity().getWorld().getEnvironment().equals(World.Environment.THE_END)
-                && !event.getEntity().isEmpty()) {
-            event.setCancelled(true);
-        }
-    }
-}
